third_party/WebKit/LayoutTests/svg/dom/transform-parser-expected.txt

CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix( 78)2	8	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.)2	9e0 795	0,- e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e,4(7- )+e(e8."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(74- 9.(+ e.0e635"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-9,).8e7e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrixx(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matri(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="matrix( )"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+e(e8.)74- 9.(+ "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.0e635(-9,).8e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotatee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotat(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="rotate( )"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+e(e8.)74- 9.(+ "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.0e635(-9,).8e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scalee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scal(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="scale( )"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+e(e8.)74- 9.(+ "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.0e635(-9,).8e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewXX(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skew(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewX( )"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+e(e8.)74- 9.(+ "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.0e635(-9,).8e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewYY(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skew(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="skewY( )"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate()"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+e(e8.)74- 9.(+ "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.0e635(-9,).8e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e4e9ee "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(49-,,4)30378.052"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e7863e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e72(	e	,49.7."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e55, "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(263,ee1)74-	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.64110,.+1."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.59e3...725-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6	-)2).180"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(,e 579	++."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(74e	-	6	-)29	7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(221e..+	+77"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate())4e	1e).-e8378."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	4+01 	e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(-6++9++ 9.3)6438"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(-,15e)(-+,- 6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(3e+)085932"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	0944.6-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(3e(-2e3e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6	5e 3e	(.	1 69	16"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(1ee5 56,(8.5	e,3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e-		 e-"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate()e0e)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(231293.16e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(()2(37-888580(9e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(18e805"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(63) 693("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(80,2)+ 4ee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(5e+59e2e5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(5	3e,9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e7	)((59	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(963.+8-e73-e		+)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(71(1(+-3382.3"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate( 27("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(88"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(9 43,+6 87.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	-	 .5997"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(5.9)557"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+65)e)0,	217"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(+	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e8+(++	e52,. 73"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(,0+	."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(-59-62159544,8,	3	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0857) 5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6,71+5e )e "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(2e5	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(1e2)7e()	90"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(9.-)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(8e759.6,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(2(65	.59e0+8-5"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.0712"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e97,e33-)6,(e0-1	+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(. 3 6.e,)57e7.	(6"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(1-7.(e(8.e-)e.) 63	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6786- +8912e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate()ee8),-886().9	--	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.,e5732."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(18-+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(3,(13e+711e9eee0.)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6	"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate()	+69"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(, 19e	)2)+0-(9)2"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(1e,--+810)5..e37"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(35 )+e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(362020()8347081e78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate()+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e-.54(+006	3141e"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e)	.1+9e.0)7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(61)2e."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(7075-5(e67)+(-7)7)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e  "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(88)(3e21"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(3e-003(4"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(682- 	19-0)4642"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(6 7"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(.9e1.e6+,e0-	20e	)9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(e+"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(96,e2	0 1e2-5)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate( 8-7(3."
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate("
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(8 78"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(2	8	2e.172	(0 .)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(	9"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate(0 795	0,- eee,"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate((7- "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translatee(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translat(0, 0)"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate"
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate "
CONSOLE ERROR: Error: Invalid value for <g> attribute transform="translate( )"
This test fuzzes the transform parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Parsed as 1 transform(s): matrix(0,0,0,0,0,0)
Parsed as 1 transform(s): rotate(0)
Parsed as 1 transform(s): rotate(0,0,0)
Parsed as 1 transform(s): scale(0)
Parsed as 1 transform(s): scale(0,0)
Parsed as 1 transform(s): skewX(0)
Parsed as 1 transform(s): skewY(0)
Parsed as 1 transform(s): translate(0)
Parsed as 1 transform(s): translate(0,0)
PASS successfullyParsed is true

TEST COMPLETE
